+++ pytsite/unittest/__init__.py
@@ -1,0 +1,7 @@
+"""
+"""
+from ._test_case import TestCase
+
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'